requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.3.11=haa95532_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
colorama=0.4.6=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
docker=7.0.0=pypi_0
flaml=2.1.2=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=pypi_0
libffi=3.4.4=hd77b12b_1
numpy=1.26.4=pypi_0
openai=1.30.1=pypi_0
openssl=3.0.13=h2bbff1b_1
packaging=24.0=pypi_0
pip=24.0=py311haa95532_0
pyautogen=0.2.27=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
python=3.11.9=he1021f5_0
python-dotenv=1.0.1=pypi_0
pywin32=306=pypi_0
regex=2024.5.15=pypi_0
requests=2.31.0=pypi_0
setuptools=69.5.1=py311haa95532_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
termcolor=2.4.0=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.14=h0416ee5_0
tqdm=4.66.4=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.1=pypi_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
wheel=0.43.0=py311haa95532_0
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
